Divide js code into several modules
